min fix, rename style
